+++ VariableBase.Mathematics/VariableBase.Mathematics/Interfaces/IMathEnvironment.cs
@@ -9,15 +9,15 @@ namespace VariableBase.Mathematics
     {
         IBasicMathAlgorithm BasicMath { get; set; }
         IPrimeAlgorithm PrimeAlgorithm { get; set; }
-        Double Base { get; }
+        Decimal Base { get; }
         ReadOnlyCollection<Char> Key { get; }
         ReadOnlyCollection<Number> KeyNumber { get; }
         Number PowerOfFirstNumber { get; }
         Number SecondNumber { get; }
-        Number ConvertToFraction(Double numberRaw, Double numeratorNumber, Double denominatorRaw);
+        Number ConvertToFraction(Decimal numberRaw, Decimal numeratorNumber, Decimal denominatorRaw);
         Boolean Equals(Object other);
         Int32 GetHashCode();
-        Double GetIndex(Char arg);
+        Decimal GetIndex(Char arg);
         Number GetNumber(Int32 zeros, Boolean isNegative = false);
         Number GetNumber(String wholeNumber, String fractionNumerator = null, String fractionDenominator = null, Boolean isNegative = false);
         Number GetNumber(String[] wholeNumberSegments, Boolean isNegative = false);
